examples/agent-runtime/agentic-frameworks/autogen/pyproject.toml
[project]
name = "autogen-agent-example"
version = "1.0.0"
description = "Microsoft AutoGen multi-agent conversation framework integration with Novita Agent Runtime example project"
authors = [
    {name = "Jason", email = "jason.zhang@novita.ai"}
]
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}

dependencies = [
    "novita-sandbox>=1.1.0b1",
    "pyautogen>=0.2.0",
    "autogen-agentchat>=0.4.0",
    "autogen-ext>=0.4.0",
    "openai>=1.50.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.black]
line-length = 100
target-version = ['py310']

[tool.ruff]
line-length = 100
target-version = "py310"

